requirements.txt
yfinance
pandas
numpy
re
colorama
scipy.stats